pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "outlook-ai"
version = "0.1.0"
description = "AI-powered Outlook.com CLI assistant using Ollama"
readme = "README.md"
authors = [{name = "Yuan Li"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "msal>=1.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
outlook-ai = "outlook_ai.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
